add responsive design, updated context
