+++ client/components/MenuPanel/menuPanel.module.scss
@@ -1,0 +1,39 @@
+h2 {
+  font-family: 'Roboto', 'Helvetica', 'Arial', sans-serif;
+}
+
+.menu {
+  //margin: 5%;
+  padding: 6%;
+}
+
+button {
+  color: rgba(0, 0, 0, 0.87);
+  padding: 6px 12px;
+  font-size: 0.875rem;
+  min-width: 64px;
+  box-sizing: border-box;
+  transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
+    box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,
+    border 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
+  font-family: 'Roboto', 'Helvetica', 'Arial', sans-serif;
+  font-weight: 500;
+  line-height: 1.75;
+  border-radius: 4px;
+  letter-spacing: 0.02857em;
+  text-transform: uppercase;
+  border: none;
+  margin: 5px;
+}
+
+button:hover {
+  background-color: rgba(0, 0, 0, 0.04);
+  text-decoration: none;
+  border: none;
+}
+
+button:focus {
+  background-color: rgba(0, 0, 0, 0.114);
+  text-decoration: none;
+  border: none;
+}
